use cookies to pass information from server to client when importing notion
